+++ lib/lottie/gui/__init__.py
@@ -1,5 +1,5 @@
-from . import import_export, tree_view, timeline_widget
+from . import import_export, tree_view, timeline_widget, console
 
 __all__ = [
-    "import_export", "tree_view", "timeline_widget"
+    "import_export", "tree_view", "timeline_widget", "console"
 ]