+++ CHANGELOG.md
@@ -2,6 +2,7 @@
  * Support for all importers / exporters in lottie GUI
  * JSON / Source editing in lottie GUI
  * SVG importer improvements
+ * Python console in lottie GUI
 
 ## 0.6.1 - 2020-04-25
  * Made scripts more user friendly